modify AutoComplete control
